requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlmodel==0.0.16
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
jinja2==3.1.3
python-dotenv==1.0.1 
openai==1.68.0
httpx
